+++ src/com/javamentor/Stepik_01_03_01/Quiz.java
@@ -2,7 +2,7 @@
  * Чтобы убедиться, что вы действительно научились пользоваться компилятором и запускать Java-программы,
  * запустите следующую программу. В качестве ответа введите то, что эта программа напечатает в консоль.
  */
-package com.javamentor.Stepik_01_03_03;
+package com.javamentor.Stepik_01_03_01;
 
 import java.security.MessageDigest;
